Finish Journal module

+++ src/modules/journal/views/EntryView.vue
@@ -1,44 +1,44 @@
 <template>
     <template v-if="entry">
-  <div  class="entry-title d-flex justify-content-between p-2">
-    <div>
-        <span class="text-success fs-3 fw-bold">{{ day }}</span>
-        <span class="mx-1 fs-3">{{ month }}</span>
-        <span class="mx-1 fs-4 fw-light">{{ yearDay }}</span>
-    </div>
+        <div class="entry-title d-flex justify-content-between p-2">
+            <div>
+                <span class="text-success fs-3 fw-bold">{{ day }}</span>
+                <span class="mx-1 fs-3">{{ month }}</span>
+                <span class="mx-1 fs-4 fw-light">{{ yearDay }}</span>
+            </div>
 
-    <div>
-        <button v-if="entry.id"
-            class="btn btn-info mx-2"
-            @click="deleteEntryId">
-            Delete
-            <i class="fa fa-trash-alt"></i>
-        </button>
+            <div>
+                <input type="file" @change="onSelectedImage" ref="imageSelector" v-show="false"
+                    accept="image/png, image/jpg, image/jpeg">
+                <button v-if="entry.id" class="btn btn-info mx-2" @click="deleteEntryId">
+                    Delete
+                    <i class="fa fa-trash-alt"></i>
+                </button>
 
-        <button class="btn btn-primary">
-            Upload picture
-            <i class="fa fa-upload"></i>
-        </button>
-    </div>
-  </div>
+                <button class="btn btn-primary" @click="selectImage">
+                    Upload picture
+                    <i class="fa fa-upload"></i>
+                </button>
+            </div>
+        </div>
 
-  <hr>
-  <div class="d-flex flex-column px-3 h-75">
-    <textarea name="" id="" cols="30" rows="10"
-            placeholder="Tell me your point of view today" 
-            v-model="entry.text"   
-    ></textarea>
-  </div>
-  <img src="https://www.robertlandscapes.com/wp-content/uploads/2014/11/landscape-322100_1280.jpg" alt="entry-picture" class="img-thumbnail">
-</template>
-  <Fab icon='fa-save'
-        @on:click="saveEntry"/>
+        <hr>
+        <div class="d-flex flex-column px-3 h-75">
+            <textarea name="" id="" cols="30" rows="10" placeholder="Tell me your point of view today"
+                v-model="entry.text"></textarea>
+        </div>
+        <img v-if="entry.picture && !localImage" :src="entry.picture" alt="entry-picture" class="img-thumbnail">
+        <img v-if="localImage" :src="localImage" alt="entry-picture" class="img-thumbnail">
+    </template>
+    <Fab icon='fa-save' @on:click="saveEntry" />
 </template>
 
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex'
-import  getTodayDate from '../components/helpers/getTodayDate'
+import Swal from 'sweetalert2'
+import getTodayDate from '../components/helpers/getTodayDate'
+import uploadImage from '../components/helpers/uploadImage'
 
 export default {
     props: {
@@ -47,9 +47,11 @@ export default {
             required: true
         }
     },
-    data (){
+    data() {
         return {
-            entry: null
+            entry: null,
+            localImage: null,
+            file: null
         }
     },
     components: {
@@ -57,15 +59,15 @@ export default {
     },
     computed: {
         ...mapGetters('journal', ['getEntriesById']),
-        day(){
+        day() {
             const { day } = getTodayDate(this.entry.date)
             return day
         },
-        month(){
+        month() {
             const { month } = getTodayDate(this.entry.date)
             return month
         },
-        yearDay(){
+        yearDay() {
             const { yearDay } = getTodayDate(this.entry.date)
             return yearDay
         }
@@ -73,39 +75,95 @@ export default {
     methods: {
         ...mapActions('journal', ['updatedEntry', 'createEntry', 'deleteEntry']),
 
-        loadEntry(){
+        loadEntry() {
             let entry
-            if(this.id === 'new') {
+            if (this.id === 'new') {
                 entry = {
                     text: 'Tell me your point of view today...',
                     date: new Date().getTime()
                 }
             } else {
                 entry = this.getEntriesById(this.id)
-                if (!entry) return this.$router.push({ name: 'no-entry'})
+                if (!entry) return this.$router.push({ name: 'no-entry' })
 
             }
             this.entry = entry
         },
-        async saveEntry(){
-            if(this.entry.id){
+        async saveEntry() {
+            new Swal({
+                title: 'Loading...',
+                allowOutsideClick: false
+            })
+            Swal.showLoading()
+
+            const picture = await uploadImage(this.file);
+            if (picture) {
+                this.entry.picture = picture;
+            }
+
+            if (this.entry.id) {
                 await this.updatedEntry(this.entry)
             } else {
                 const id = await this.createEntry(this.entry)
-                this.$router.push({ name: 'entry', params: { id }})
+                this.$router.push({ name: 'entry', params: { id } })
             }
-            
+            this.file = null
+            Swal.fire({
+                icon: 'success',
+                title: 'Save',
+                text: 'This entry saved successfully',
+                timer: 3000
+            })
+            this.$router.push({ name: 'no-entry'})
         },
         async deleteEntryId() {
-            await this.deleteEntry(this.entry.id)
-            this.$router.push({ name: 'no-entry'})
+            await Swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then(async (result) => {
+                if (result.isConfirmed) {
+                    new Swal({
+                        title: 'Loading...',
+                        allowOutsideClick: false
+                    })
+                    Swal.showLoading()
+                    await this.deleteEntry(this.entry.id)
+                    this.$router.push({ name: 'no-entry' })
+                    Swal.fire(
+                        'Deleted!',
+                        'Your entry has been deleted.',
+                        'success'
+                    )
+                }
+            })
+        },
+        onSelectedImage(e) {
+            const file = e.target.files[0]
+            if (!file) {
+                this.localImage = null
+                this.file = null
+                return
+            }
+
+            this.file = file
+            const fr = new FileReader()
+            fr.onload = () => this.localImage = fr.result
+            fr.readAsDataURL(file)
+        },
+        selectImage() {
+            this.$refs.imageSelector.click()
         }
     },
     created() {
         this.loadEntry()
     },
-    watch:{
-        id(){
+    watch: {
+        id() {
             this.loadEntry()
         }
     }
@@ -113,14 +171,13 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-
 textarea {
     font-size: 20px;
     border: none;
     height: 92%;
     background-color: aliceblue;
 
-    &:focus{
+    &:focus {
         outline: none;
     }
 }
@@ -132,5 +189,4 @@ img {
     right: 20px;
     box-shadow: 0px 5px 10px rgba($color: #000000, $alpha: 0.2);
 }
-
 </style>
